+++ contracts/btc-finality/src/migrations/mod.rs
@@ -1,0 +1,3 @@
+pub mod v1_0_0_rc_0_to_v1_0_0_rc_1;
+
+pub use v1_0_0_rc_0_to_v1_0_0_rc_1::migrate_config as migrate_config_v1_0_0_rc_0_to_v1_0_0_rc_1;